+++ app/src/ui/containers/sandwich-machine.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { Button } from './../components/Button';
-import { getIngredients } from './../../state/ingredients/ingredients.reducer';
-import { checkIfSandwich } from './../../state/ingredients/ingredients.reducer';
+import { getIngredients } from './../../state/ingredients/ingredients.actions';
+import { checkIfSandwich } from './../../state/ingredients/ingredients.actions';
 
 export const SandwichMachine = connect(
   (state: any) => ({
@@ -18,8 +18,8 @@ export const SandwichMachine = connect(
       <Button onPress={requestIngredients} value={'Get Ingredients'} />
       <Button onPress={checkSandwich} value={'Check if sandwich'} />
       <p>
-        { ingredients[0] ? ingredients[0].name : '' }
+        {ingredients[0] ? ingredients[0].name : ''}
       </p>
-      <h1>{ isSandwich ? 'IS SANDWICH!' : 'IS NO SANDWICH!' } </h1>
+      <h1>{isSandwich ? 'IS SANDWICH!' : 'IS NO SANDWICH!'} </h1>
     </div>
   ));
